include statistics in OuquTp transpiler results
